# contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Internal Audience & Campaign Control API
  version: 0.8.28
  description: >
    Internal-only API for organisation records, offline consent reviews, secure audience imports,
    governed cohort definitions and controlled campaign lifecycle management.
servers:
  - url: /api/v1
security:
  - internalSession: []
paths:
...
  /admin/opt-out-policies:
    get:
      operationId: listOptOutPolicies
      summary: List governed effective-dated opt-out policies
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque continuation token returned by the previous page. }
      responses:
        '200':
          description: Policy list
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
    post:
      operationId: createOptOutPolicy
      summary: Create a draft exact-match opt-out policy
      responses:
        '201': { description: Draft policy created }
  /admin/opt-out-policies/{id}/submit:
    post:
      operationId: submitOptOutPolicy
      summary: Submit an opt-out policy for independent approval
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Policy awaiting approval }
  /admin/opt-out-policies/{id}/decision:
    post:
      operationId: decideOptOutPolicy
      summary: Approve or reject a submitted opt-out policy
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Decision recorded }
...
components:
  securitySchemes:
    internalSession:
      type: apiKey
      in: cookie
      name: campaign_session
...